+++ draft-packages/title-block-zen/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.25](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@5.3.24...@kaizen/draft-title-block-zen@5.3.25) (2022-03-16)
+
+**Note:** Version bump only for package @kaizen/draft-title-block-zen
+
+
+
+
+
 ## [5.3.24](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@5.3.23...@kaizen/draft-title-block-zen@5.3.24) (2022-03-16)
 
 **Note:** Version bump only for package @kaizen/draft-title-block-zen